Updated for buy,rent,commercial pages

+++ api/controllers/user.controller.js
@@ -68,10 +68,15 @@ export const getUsers = async (req, res, next) => {
 
 export const deleteUser = async (req, res, next) => {
   try {
-    const deletedUser = await User.findByIdAndDelete(req.params.id);
+    const deletedUser = await User.findById(req.params.id);
+
     if (!deletedUser) {
       return next(errorHandler(404, "User not found"));
     }
+  if (deletedUser.email === 'mohamed@agcoproperties.com') {
+      return next(errorHandler(403, "System administrator account cannot be deleted"));
+    }
+    await User.findByIdAndDelete(req.params.id);
     res.status(200).json({ message: "User deleted successfully" });
   } catch (error) {
     next(error);
